requirements.txt
streamlit==1.32.2
torch==2.3.1
torchvision==0.18.1
pillow==10.2.0
numpy==1.26.4
pandas==2.2.1
altair==5.2.0
reportlab==4.1.0